Update for 1.0.3 release (#1475)
* Fix TransferPacker, add player transfer api

* Transferring. No more raw api in class Player!

* Release

* Remove Transferring class, readd transfer api in Player class

* Improve player transfer api

* Add player device model api

+++ src/main/java/cn/nukkit/network/protocol/TransferPacket.java
@@ -5,7 +5,7 @@ public class TransferPacket extends DataPacket {
     public static final byte NETWORK_ID = ProtocolInfo.TRANSFER_PACKET;
     
     public String address; // Server address
-    public short port = 19132; // Server port
+    public int port = 19132; // Server port
     
     @Override
     public void decode() {
